refactor: 어드민 패키지 정리
