+++ backend/src/main/java/com/festago/admin/application/AdminArtistV1QueryService.java
@@ -1,6 +1,6 @@
-package com.festago.artist.application;
+package com.festago.admin.application;
 
-import com.festago.admin.dto.ArtistV1Response;
+import com.festago.admin.dto.artist.ArtistV1Response;
 import com.festago.artist.domain.Artist;
 import com.festago.artist.repository.ArtistRepository;
 import java.util.List;
@@ -8,10 +8,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+// TODO 페이징 적용
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class ArtistV1QueryService {
+public class AdminArtistV1QueryService {
 
     private final ArtistRepository artistRepository;
