+++ backend/src/main/java/com/festago/admin/application/AdminSchoolV1QueryService.java
@@ -1,6 +1,6 @@
 package com.festago.admin.application;
 
-import com.festago.admin.dto.AdminSchoolV1Response;
+import com.festago.admin.dto.school.AdminSchoolV1Response;
 import com.festago.admin.repository.AdminSchoolV1QueryDslRepository;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
