+++ backend/src/main/java/com/festago/admin/dto/artist/ArtistCreateRequest.java
@@ -1,8 +1,8 @@
-package com.festago.admin.dto;
+package com.festago.admin.dto.artist;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record ArtistUpdateRequest(
+public record ArtistCreateRequest(
     @NotBlank
     String name,
     @NotBlank
