+++ backend/src/main/java/com/festago/admin/dto/festival/AdminFestivalDetailV1Response.java
@@ -1,4 +1,4 @@
-package com.festago.admin.dto;
+package com.festago.admin.dto.festival;
 
 import com.querydsl.core.annotations.QueryProjection;
 import java.time.LocalDate;
