+++ backend/src/main/java/com/festago/admin/presentation/v1/AdminFestivalV1Controller.java
@@ -1,10 +1,10 @@
 package com.festago.admin.presentation.v1;
 
 import com.festago.admin.application.AdminFestivalV1QueryService;
-import com.festago.admin.dto.AdminFestivalDetailV1Response;
-import com.festago.admin.dto.AdminFestivalV1Response;
-import com.festago.admin.dto.FestivalV1CreateRequest;
-import com.festago.admin.dto.FestivalV1UpdateRequest;
+import com.festago.admin.dto.festival.AdminFestivalDetailV1Response;
+import com.festago.admin.dto.festival.AdminFestivalV1Response;
+import com.festago.admin.dto.festival.FestivalV1CreateRequest;
+import com.festago.admin.dto.festival.FestivalV1UpdateRequest;
 import com.festago.common.aop.ValidPageable;
 import com.festago.common.querydsl.SearchCondition;
 import com.festago.festival.application.command.FestivalCommandFacadeService;
