+++ backend/src/test/java/com/festago/admin/application/integration/AdminSchoolV1QueryServiceIntegrationTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import com.festago.admin.application.AdminSchoolV1QueryService;
-import com.festago.admin.dto.AdminSchoolV1Response;
+import com.festago.admin.dto.school.AdminSchoolV1Response;
 import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
 import com.festago.common.querydsl.SearchCondition;
